// apps/helpdesk/src/auto_qa/types.ts
import type { ReactNode } from 'react'

import type { TicketQAScoreDimension } from '@gorgias/helpdesk-queries'

export type DimensionConfig = {
    autoExpandThreshold?: number
    label: string
    options: DimensionOption[]
    prefix?: ReactNode
    tooltip?: string
    placeholder?: string
    scorePlaceholder?: string
}

export type DimensionOption = {
    label: string
    value: number
}

export type DimensionSummary = Pick<
    TicketQAScoreDimension,
    'explanation' | 'prediction'
>